# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: getStudents
      tags: [students]
      summary: List students
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [active, inactive, graduated] }
      responses:
        "200":
          description: Student list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Student" }
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a student
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StudentInput" }
      responses:
        "201":
          description: Created student
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Student" }
  /students/{id}:
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StudentUpdate" }
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Student" }
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: number }
      responses:
        "204": { description: Student deleted }
  /teachers:
    get:
      operationId: getTeachers
      tags: [teachers]
      summary: List teachers
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [active, inactive] }
      responses:
        "200":
          description: Teacher list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Teacher" }
    post:
      operationId: createTeacher
      tags: [teachers]
      summary: Create a teacher
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TeacherInput" }
      responses:
        "201":
          description: Created teacher
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Teacher" }
  /teachers/{id}:
    patch:
      operationId: updateTeacher
      tags: [teachers]
      summary: Update a teacher
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TeacherUpdate" }
      responses:
        "200":
          description: Updated teacher
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Teacher" }
    delete:
      operationId: deleteTeacher
      tags: [teachers]
      summary: Delete a teacher
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: number }
      responses:
        "204": { description: Teacher deleted }
  /employees:
    get:
      operationId: getEmployees
      tags: [employees]
      summary: List employees
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [active, inactive] }
      responses:
        "200":
          description: Employee list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Employee" }
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create an employee
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeInput" }
      responses:
        "201":
          description: Created employee
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Employee" }
  /employees/{id}:
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeUpdate" }
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Employee" }
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete an employee
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: number }
      responses:
        "204": { description: Employee deleted }
  /library/books:
    get:
      operationId: getBooks
      tags: [library]
      summary: List library books
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
      responses:
        "200":
          description: Book list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Book" }
    post:
      operationId: createBook
      tags: [library]
      summary: Add a library book
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookInput" }
      responses:
        "201":
          description: Created book
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Book" }
  /library/books/{id}:
    patch:
      operationId: updateBook
      tags: [library]
      summary: Update a library book
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookUpdate" }
      responses:
        "200":
          description: Updated book
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Book" }
    delete:
      operationId: deleteBook
      tags: [library]
      summary: Delete a library book
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: number }
      responses:
        "204": { description: Book deleted }
  /library/borrows:
    get:
      operationId: getBorrows
      tags: [library]
      summary: List borrows
      parameters:
        - in: query
          name: active
          schema: { type: boolean }
      responses:
        "200":
          description: Borrow list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Borrow" }
    post:
      operationId: createBorrow
      tags: [library]
      summary: Borrow a book for a student, teacher, or employee
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BorrowInput" }
      responses:
        "201":
          description: Created borrow
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Borrow" }
  /library/books/{id}/condition:
    patch:
      operationId: markBookCondition
      tags: [library]
      summary: Mark one copy of a book as lost or damaged, or restore a copy
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookConditionInput" }
      responses:
        "200":
          description: Updated book
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Book" }
        "404": { description: Book not found }
        "409": { description: No copy available for the requested action }
  /library/borrows/{id}/return:
    patch:
      operationId: returnBorrow
      tags: [library]
      summary: Mark a borrow as returned
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: number }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookReturnInput" }
      responses:
        "200":
          description: Returned borrow
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Borrow" }
        "404": { description: Borrow not found }
        "409": { description: Borrow already returned }
...
components:
  schemas:
...
    Student:
      type: object
      required: [id, fullName, fullNameArabic, gender, studentNumber, nationalId, grade, className, status, enrollmentDate]
      properties:
        id: { type: number }
        fullName: { type: string }
        fullNameArabic: { type: string }
        gender: { type: string, enum: [male, female] }
        studentNumber: { type: string }
        nationalId: { type: string }
        grade: { type: string }
        className: { type: string }
        guardianName: { type: string }
        guardianPhone: { type: string }
        status: { type: string, enum: [active, inactive, graduated] }
        enrollmentDate: { type: string, format: date }
    StudentInput:
      type: object
      required: [fullName, fullNameArabic, gender, studentNumber, nationalId, grade, className, enrollmentDate]
      properties:
        fullName: { type: string, minLength: 1 }
        fullNameArabic: { type: string, minLength: 1 }
        gender: { type: string, enum: [male, female] }
        studentNumber: { type: string, minLength: 1 }
        nationalId: { type: string, minLength: 1 }
        grade: { type: string, minLength: 1 }
        className: { type: string, minLength: 1 }
        guardianName: { type: string }
        guardianPhone: { type: string }
        enrollmentDate: { type: string, format: date }
    StudentUpdate:
      allOf:
        - $ref: "#/components/schemas/StudentInput"
    Teacher:
      type: object
      required: [id, fullName, fullNameArabic, name, surname, username, englishName, employeeCode, nationalId, nationality, gender, maritalStatus, religion, phone, email, address, area, country, height, weight, branch, academicLevel, subject, weeklyClasses, isEmployee, status]
      properties:
        id: { type: number }
        fullName: { type: string }
        fullNameArabic: { type: string }
        name: { type: string }
        surname: { type: string }
        username: { type: string }
        englishName: { type: string }
        employeeCode: { type: string }
        nationalId: { type: string }
        nationality: { type: string }
        gender: { type: string }
        maritalStatus: { type: string }
        religion: { type: string }
        phone: { type: string }
        email: { type: string }
        address: { type: string }
        area: { type: string }
        country: { type: string }
        height: { type: number }
        weight: { type: number }
        branch: { type: string }
        academicLevel: { type: string }
        subject: { type: string }
        weeklyClasses: { type: number }
        isEmployee: { type: boolean }
        status: { type: string, enum: [active, inactive] }
...
    EmployeeUpdate:
      allOf:
        - $ref: "#/components/schemas/EmployeeInput"
...
    BorrowInput:
      type: object
      required: [bookId, borrowerType, borrowerId]
      properties:
        bookId: { type: number }
        borrowerType: { type: string, enum: [student, teacher, employee] }
        borrowerId: { type: number }
        dueDate: { type: string, format: date }
...